requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
aiosmtplib==4.0.0
email-validator==2.2.0
python-dotenv==1.1.0